Fixes for 2nd task "Components"
