+++ project/src/components/app/app.tsx
@@ -1,7 +1,7 @@
-import Main, {PromoFilmData} from '../../pages/main/main';
+import Main, {MainProps} from '../../pages/main/main';
 
-function App(props: PromoFilmData): JSX.Element {
-  return Main(props);
+function App(props: MainProps): JSX.Element {
+  return <Main promoData={props.promoData} films={props.films} genres={props.genres}/>;
 }
 
 export default App;
