+++ project/src/components/page-footer/page-footer.tsx
@@ -1,0 +1,15 @@
+import Logo from "../logo/logo";
+
+function Footer() {
+  return (
+    <footer className="page-footer">
+      <Logo className={"logo__link logo__link--light"}/>
+
+      <div className="copyright">
+        <p>© 2019 What to watch Ltd.</p>
+      </div>
+    </footer>
+  );
+}
+
+export default Footer;
