+++ project/src/types/promo-film-data.tsx
@@ -1,0 +1,7 @@
+export type PromoFilmData = {
+  title: string;
+  genre: string;
+  releaseYear: number;
+  bgImagePath: string;
+  imagePath: string;
+}
